Reformat the user data so it is more readable
