+++ HitachiEIP/HitachiEIP/Data.cs
@@ -291,249 +291,6 @@ namespace HitachiEIP {
 
          RFS.WriteLine();
       }
-      #endregion
-
-      #region Conversion Tables
-
-      // Attribute DropDown conversion
-      public static string[][] DropDowns = new string[][] {
-         new string[] { },                                            // 0 - Just decimal values
-         new string[] { "Disable", "Enable" },                        // 1 - Enable and disable
-         new string[] { "Disable", "Space Fill", "Character Fill" },  // 2 - Disable, space fill, character fill
-         new string[] { "TwentyFour Hour", "Twelve Hour" },           // 3 - 12 & 24 hour
-         new string[] { "Current Time", "Stop Clock" },               // 4 - Current time or stop clock
-         new string[] { "Off Line", "On Line" },                      // 5 - Offline/Online
-         new string[] { "None", "Signal 1", "Signal 2" },             // 6 - None, Signal 1, Signal 2
-         new string[] { "Up", "Down" },                               // 7 - Up/Down
-         new string[] { "None", "5X5", "5X7" },                       // 8 - Readable Code 5X5 or 5X7
-         new string[] { "not used", "code 39", "ITF", "NW-7", "EAN-13", "DM8x32", "DM16x16", "DM16x36",
-                        "DM16x48", "DM18x18", "DM20x20", "DM22x22", "DM24x24", "Code 128 (Code set B)",
-                        "Code 128 (Code set C)", "UPC-A", "UPC-E", "EAN-8", "QR21x21", "QR25x25", "QR29x29",
-                        "GS1 DataBar (Limited)", "GS1 DataBar (Omnidirectional)", "GS1 DataBar (Stacked)", "DM14x14", },
-                                                                      // 9 - Barcode Types
-         new string[] { "Normal", "Reverse" },                        // 10 - Normal/reverse
-         new string[] { "M 15%", "Q 25%" },                           // 11 - M 15%, Q 25%
-         new string[] { "Edit Message", "Print Format" },             // 12 - Edit/Print
-         new string[] { "From Yesterday", "From Today" },             // 13 - From Yesterday/Today
-         new string[] { "4x5", "5x5", "5x8(5x7)", "9x8(9x7)", "7x10", "10x12", "12x16", "18x24", "24x32",
-                        "11x11", "5x3(Chimney)", "5x5(Chimney)", "7x5(Chimney)", "QR33", "30x40", "36x48"  },
-                                                                      // 14 - Font Types
-         new string[] { "Normal/Forward", "Normal/Reverse",
-                         "Inverted/Forward", "Inverted/Reverse",},    // 15 - Orientation
-         new string[] { "None", "Encoder", "Auto" },                  // 16 - Product speed matching
-         new string[] { "HM", "NM", "QM", "SM" },                     // 17 - High Speed Print
-         new string[] { "Time Setup", "Until End of Print" },         // 18 - Target Sensor Filter
-         new string[] { "4x5", "5x5", "5x8(5x7)", "9x8(9x7)", "7x10", "10x12", "12x16", "18x24", "24x32",
-                        "11x11", "5x3(Chimney)", "5x5(Chimney)", "7x5(Chimney)", "30x40", "36x48"  },
-                                                                      // 19 - User Pattern Font Types
-      };
-
-      //// Class Codes to Attributes
-      //public static Type[] ClassCodeAttributes = new Type[] {
-      //      typeof(ccPDM),   // 0x66 Print data management function
-      //      typeof(ccPF),    // 0x67 Print format function
-      //      typeof(ccPS),    // 0x68 Print specification function
-      //      typeof(ccCal),   // 0x69 Calendar function
-      //      typeof(ccUP),    // 0x6B User pattern function
-      //      typeof(ccSR),    // 0x6C Substitution rules function
-      //      typeof(ccES),    // 0x71 Enviroment setting function
-      //      typeof(ccUI),    // 0x73 Unit Information function
-      //      typeof(ccOM),    // 0x74 Operation management function
-      //      typeof(ccIJP),   // 0x75 IJP operation function
-      //      typeof(ccCount), // 0x79 Count function
-      //      typeof(ccIDX),   // 0x7A Index function
-      //   };
-
-      //// Class Names
-      //public static string[] ClassNames = Enum.GetNames(typeof(ClassCode));
-
-      //// Class Codes
-      //public static ClassCode[] ClassCodes = (ClassCode[])Enum.GetValues(typeof(ClassCode));
-
-      // Class Codes to Data Tables Conversion
-      private static int[][][] ClassCodeData = new int[][][] {
-            PrintDataManagement,           // 0x66 Print data management function
-            PrintFormat,                   // 0x67 Print format function
-            PrintSpecification,            // 0x68 Print specification function
-            Calendar,                      // 0x69 Calendar function
-            UserPattern,                   // 0x6B User pattern function
-            SubstitutionRules,             // 0x6C Substitution rules function
-            EnviromentSetting,             // 0x71 Enviroment setting function
-            UnitInformation,               // 0x73 Unit Information function
-            OperationManagement,           // 0x74 Operation management function
-            IJPOperation,                  // 0x75 IJP operation function
-            Count,                         // 0x79 Count function
-            Index,                         // 0x7A Index function
-         };
-
-      #endregion
-
-      #region Service Routines
-
-      //// Lookup for getting attributes associated with a Class/Function
-      //public static Dictionary<ClassCode, byte, AttrData> AttrDict;
-
-      // Get attribute data for an arbitrary class/attribute
-      //public static AttrData GetAttrData(ClassCode Class, byte attr) {
-      //   int[][] tab = ClassCodeData[Array.IndexOf(ClassCodes, Class)];
-      //   for (int j = 0; j < tab.Length; j++) {
-      //      if ((byte)tab[j][0] == attr) {
-      //         return new AttrData(tab[j]);
-      //      }
-      //   }
-      //   return null;
-      //}
-
-      //// Build the Attribute Dictionary
-      //public static void BuildAttributeDictionary() {
-      //   if (AttrDict == null) {
-      //      AttrDict = new Dictionary<ClassCode, byte, AttrData>();
-      //      for (int i = 0; i < ClassCodes.Length; i++) {
-      //         int[] ClassAttr = (int[])ClassCodeAttributes[i].GetEnumValues();
-      //         for (int j = 0; j < ClassAttr.Length; j++) {
-      //            AttrDict.Add(ClassCodes[i], (byte)ClassAttr[j], DataII.GetAttrData(ClassCodes[i], (Byte)ClassAttr[j]));
-      //         }
-      //      }
-      //   }
-      //}
-
-      #endregion
-
-   }
-
-   public class AttrData {
-
-      #region Properties and Constructor
-
-      // Class Code Attributes = {
-      //   [0] = Value
-      //   [1] = Set Available
-      //   [2] = Get Available
-      //   [3] = Service Available
-      //   [4] = Data Length
-      //   [5] = Format
-      //   [6] = Min Value
-      //   [7] = Max Value
-      //   [8] = AlphaSortOrder 
-      //   [9] = Ignore due to error
-      //   [10] = Drop Down
-      //   [11] = Data Length
-      //   [12] = Format
-      //   [13] = Min Value
-      //   [14] = Max Value
-
-      public byte Val { get; set; } = 0;
-      public bool HasSet { get; set; } = false;
-      public bool HasGet { get; set; } = false;
-      public bool HasService { get; set; } = false;
-      public int Order { get; set; } = 0;
-      public bool Ignore { get; set; } = false;
-      public int DropDown { get; set; } = -1;
-
-      public Prop Data { get; set; }
-      public Prop Get { get; set; }
-      public Prop Set { get; set; }
-      public Prop Service { get; set; }
-
-      public AttrData() {
-
-      }
-
-      public AttrData(int[] values) {
-         Val = (byte)values[0];
-         HasSet = values[1] > 0;
-         HasGet = values[2] > 0;
-         HasService = values[3] > 0;
-         Order = values[8] - 1;
-         Ignore = values[9] > 0;
-         DropDown = values[10];
-
-         Data = new Prop(values[4], (DataFormats)values[5], values[6], values[7]);
-         if (HasSet) {
-            Set = Data;
-         }
-         if (values.Length == 11) {
-            if (HasGet) {
-               Get = new Prop(0, DataFormats.Decimal, 0, 0);
-            } else if (HasService) {
-               Service = Get = new Prop(0, DataFormats.Decimal, 0, 0);
-            }
-         } else {
-            if (HasGet) {
-               Get = new Prop(values[11], (DataFormats)values[12], values[13], values[14]);
-            } else if (HasService) {
-               Service = new Prop(values[11], (DataFormats)values[12], values[13], values[14]);
-            }
-         }
-
-      }
-
-      public AttrData(byte Val, GSS acc, bool Ignore, int Order, Prop data, Prop data2 = null) {
-         this.Val = Val;
-         this.HasSet = acc == GSS.Set || acc == GSS.GetSet;
-         this.HasGet = acc == GSS.Get || acc == GSS.GetSet;
-         this.HasService = acc == GSS.Service;
-         this.Ignore = Ignore;
-         this.Order = Order;
-
-         this.Data = Data;
-         if (HasSet) {
-            Set = Data;
-         }
-         if (data2 != null) {
-            if (HasGet) {
-               Get = data2;
-            } else if (HasService) {
-               Service = data2;
-            }
-         } else {
-            if (HasGet) {
-               Get = new Prop(0, DataFormats.Decimal, 0, 0);
-            } else if (HasService) {
-               Service = Get = new Prop(0, DataFormats.Decimal, 0, 0);
-            }
-         }
-      }
-
-      #endregion
-
-   }
-
-   public class Dictionary<TKey1, TKey2, TValue>
-      : Dictionary<Tuple<TKey1, TKey2>, TValue>, IDictionary<Tuple<TKey1, TKey2>, TValue> {
-
-      #region Constructor and methods
-
-      public TValue this[TKey1 key1, TKey2 key2] {
-         get { return base[Tuple.Create(key1, key2)]; }
-         set { base[Tuple.Create(key1, key2)] = value; }
-      }
-
-      public void Add(TKey1 key1, TKey2 key2, TValue value) {
-         base.Add(Tuple.Create(key1, key2), value);
-      }
-
-      #endregion
-
-   }
-
-   public class Prop {
-
-      #region Constructors, properties and methods
-
-      public int Len { get; set; }
-      public DataFormats Fmt { get; set; }
-      public long Min { get; set; }
-      public long Max { get; set; }
-      public fmtDD DropDown { get; set; }
-
-      public Prop(int Len, DataFormats Fmt, long Min, long Max, fmtDD DropDown = fmtDD.None) {
-         this.Len = Len;
-         this.Fmt = Fmt;
-         this.Min = Min;
-         this.Max = Max;
-         this.DropDown = DropDown;
-      }
 
       #endregion
